refactor: migrate financial data types to BigDecimal for improved precision
- Integrate `com.ionspin.kotlin:bignum` library for multiplatform decimal arithmetic
- Implement `BigDecimalSerializer` to support Kotlinx Serialization for high-precision values
- Refactor `Account`, `Budget`, `Transaction`, and `RecurringBill` domain and data models to use `BigDecimal` instead of `Double`
- Update M-PESA and Equity SMS parsers to handle monetary amounts as `BigDecimal`
- Overhaul `DoubleExt.kt` with `BigDecimal` extensions for precise currency formatting and rounding
- Update UI components and ViewModels across Account, Budget, Transaction, and Summary features to support `BigDecimal` calculations
- Refactor validation use cases to utilize `BigDecimal` comparisons and parsing logic
- Update `NotificationService` and `AndroidNotificationService` to support precise amount formatting in alerts and reminders

+++ shared/src/commonMain/kotlin/com/fintrack/shared/feature/account/domain/model/Account.kt
@@ -1,5 +1,6 @@
 package com.fintrack.shared.feature.account.domain.model
 
+import com.ionspin.kotlin.bignum.decimal.BigDecimal
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlin.time.Instant
@@ -7,9 +8,9 @@ import kotlin.time.Instant
 data class Account(
     val id: String,
     val name: String,
-    val balance: Double? = null,
-    val income: Double? = null,
-    val expense: Double? = null,
+    val balance: BigDecimal? = null,
+    val income: BigDecimal? = null,
+    val expense: BigDecimal? = null,
     val isDefault: Boolean = false,
     val type: AccountType = AccountType.GENERAL,
     val createdAt: Instant? = null,
