+++ shared/src/commonMain/kotlin/com/fintrack/shared/feature/summary/domain/model/ProfileMetrics.kt
@@ -1,9 +1,11 @@
 package com.fintrack.shared.feature.summary.domain.model
 
+import com.ionspin.kotlin.bignum.decimal.BigDecimal
+
 data class ProfileMetrics(
     val name: String,
     val email: String,
-    val netWorth: Double,
-    val savingsRate: Double?,
-    val essentialSpendRatio: Double?
+    val netWorth: BigDecimal,
+    val savingsRate: BigDecimal?,
+    val essentialSpendRatio: BigDecimal?
 )
